fix(Falsy values are now properly returned): Now only undefined will trigger the fallback to default
When a falsy value (null,0,false,'') was found during a traverse, then a fallback to default values
was performed. This lead to some unexpected behaviors where you expected a value to be false and
instead a fallback to the default values was tried, which returns undefined if no default value is
found. This was an unexpected an unwanted behavior

BREAKING CHANGE: If no path or alternatives array is provided when a function processor is provided,
then null will be returned. Previously the node was leave with an object with an undefined processor
property, which was not very useful